ISSUE #3307 - rename placeholderCard file
