+++ frontend/src/v5/ui/components/shared/linkCard/teamspaceCard/teamspacePlaceholderCard/teamspacePlaceholderCard.component.tsx
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { ImagePlaceholder, ListItem, TextPlaceholder } from './placeholderCard.styles';
+import { ImagePlaceholder, ListItem, TextPlaceholder } from './teamspacePlaceholderCard.styles';
 import { StyledCard } from '../../linkCard.styles';
 
 export const TeamspacePlaceholderCard = (): JSX.Element => (
