+++ frontend/src/v5/ui/components/shared/linkCard/teamspaceCard/teamspacePlaceholderCard/teamspacePlaceholderCard.styles.ts
